com.unity.learn.iet-framework@0.4.0-preview.1 ## [0.4.0] - 2020-06-02 ###  Changed - Refactor `TutorialWindow` to use UI Toolkit instead of IMGUI. - Breaking change: `TutorialContainer` and `TutorialParagraph` assets refactored without backward-compatibility. - Analytics: using `EditorAnalytics` instead of `UsabilityAnalytics` as the API for all events. - `TutorialPage`: instead of arbitrary set of `TutorialParagraph`s, the page has a fixed set of fields. `TutorialParagraph` will be deprecated in the near future. ### Added - `TutorialPageEditor`: a new simplified authoring view. - Preliminary localization support. - Analytics: added an event for external reference (e.g. link cards) impressions
## [0.3.0] - 2020-03-25
### Changed
 - Raised the required Unity version to 2019.3.
 - Removed flexible spaces surrounding a video paragraph.
### Fixed
 - UI flickering on macOS and Unity 2019.3 when a tutorial page had a video.

+++ Framework/Interactive Tutorials/Editor/Property Drawers/CollectionWrapperDrawer.cs
@@ -41,37 +41,37 @@ namespace Unity.InteractiveTutorials
                 EditorGUI.PropertyField(rect, reorderableList.serializedProperty.GetArrayElementAtIndex(index), true);
             label = label != null ? new GUIContent(label) : new GUIContent(property.displayName);
             reorderableList.drawHeaderCallback = delegate(Rect rect) {
-                    int oldIndent = EditorGUI.indentLevel;
-                    EditorGUI.indentLevel = 0;
-                    EditorGUI.LabelField(rect, label);
-                    EditorGUI.indentLevel = oldIndent;
-                };
+                int oldIndent = EditorGUI.indentLevel;
+                EditorGUI.indentLevel = 0;
+                EditorGUI.LabelField(rect, label);
+                EditorGUI.indentLevel = oldIndent;
+            };
             reorderableList.elementHeightCallback = delegate(int index)
-                {
-                    return EditorGUI.GetPropertyHeight(reorderableList.serializedProperty.GetArrayElementAtIndex(index)) +
-                        EditorGUIUtility.standardVerticalSpacing * 2f + 1f;
-                };
+            {
+                return EditorGUI.GetPropertyHeight(reorderableList.serializedProperty.GetArrayElementAtIndex(index)) +
+                    EditorGUIUtility.standardVerticalSpacing * 2f + 1f;
+            };
             reorderableList.onAddCallback = delegate(ReorderableList lst) {
-                    ++lst.serializedProperty.arraySize;
-                    lst.serializedProperty.serializedObject.ApplyModifiedProperties();
-                };
+                ++lst.serializedProperty.arraySize;
+                lst.serializedProperty.serializedObject.ApplyModifiedProperties();
+            };
             reorderableList.onRemoveCallback = delegate(ReorderableList lst) {
-                    int index = lst.index;
-                    if (index >= lst.serializedProperty.arraySize)
-                    {
-                        return;
-                    }
-                    SerializedProperty element = lst.serializedProperty.GetArrayElementAtIndex(index);
-                    if (
-                        element.propertyType == SerializedPropertyType.ObjectReference &&
-                        element.objectReferenceValue != null
-                        )
-                    {
-                        lst.serializedProperty.DeleteArrayElementAtIndex(index);
-                    }
+                int index = lst.index;
+                if (index >= lst.serializedProperty.arraySize)
+                {
+                    return;
+                }
+                SerializedProperty element = lst.serializedProperty.GetArrayElementAtIndex(index);
+                if (
+                    element.propertyType == SerializedPropertyType.ObjectReference &&
+                    element.objectReferenceValue != null
+                )
+                {
                     lst.serializedProperty.DeleteArrayElementAtIndex(index);
-                    lst.serializedProperty.serializedObject.ApplyModifiedProperties();
-                };
+                }
+                lst.serializedProperty.DeleteArrayElementAtIndex(index);
+                lst.serializedProperty.serializedObject.ApplyModifiedProperties();
+            };
             m_Widgets[key] = reorderableList;
             OnReorderableListCreated(reorderableList);
             return reorderableList;
